Add files via upload
The two png files are used for the GUI when running the script. Please make sure they are in the same folder as the .py file. The code is actually a simple heat-transfer calculation using iterations. The coefficients for the heat transfer (h_g for the combustion gasses and h_c for the cooling agent are actually derived using publically-available information on the Raptor engine). The wall of the cylindrical combustion chamber is made of two materials - inner and outer. You can choose the thickness and the thermal conductivity thereof. The code calculates the temperature on the inner material's surface, on the outer material's surface, and at the point of contact between the two material layers (inner and outer).